requirements.txt
absl-py==2.1.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
autoflake==2.3.1
black==24.10.0
Brotli @ file:///C:/b/abs_3d36mno480/croot/brotli-split_1714483178642/work
certifi @ file:///C:/b/abs_1fw_exq1si/croot/certifi_1725551736618/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
cleverhans==4.0.0
click==8.1.7
cloudpickle==3.1.0
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
contourpy @ file:///D:/bld/contourpy_1731428338786/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
debugpy @ file:///D:/bld/debugpy_1731044910913/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
dm-tree==0.1.8
eagerpy==0.30.0
easydict==1.13
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1720869315914/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1725214404607/work
filelock @ file:///C:/b/abs_f2gie28u58/croot/filelock_1700591233643/work
flake8==7.1.1
fonttools @ file:///C:/b/abs_f47gnfqnx0/croot/fonttools_1713551644747/work
foolbox==3.3.4
fsspec==2024.2.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
gmpy2 @ file:///C:/ci_311/gmpy2_1677743390134/work
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1726082825846/work
ipykernel @ file:///D:/bld/ipykernel_1719845595208/work
ipython @ file:///D:/bld/ipython_1729866374643/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2 @ file:///C:/b/abs_ddyeze2h6_/croot/jinja2_1730903330414/work
joblib @ file:///C:/b/abs_f4b98l6lgk/croot/joblib_1718217224240/work
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1726610684920/work
jupyter_core @ file:///D:/bld/jupyter_core_1727163532151/work
kiwisolver @ file:///C:/ci_311/kiwisolver_1676431979301/work
MarkupSafe @ file:///C:/b/abs_ecfdqh67b_/croot/markupsafe_1704206030535/work
matplotlib==3.9.1
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1713250518406/work
mccabe==0.7.0
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
mnist==0.2.2
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
mypy-extensions==1.0.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx @ file:///C:/b/abs_3bxnu56g9d/croot/networkx_1717597507456/work
nose==1.3.7
numpy==1.26.4
onnx==1.17.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1718189413536/work
pandas==2.2.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1712320355065/work
pathspec==0.12.1
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow @ file:///C:/b/abs_32o8er3uqp/croot/pillow_1721059447598/work
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1726613481435/work
ply==3.11
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1727341649933/work
protobuf==5.29.0
psutil @ file:///D:/bld/psutil_1729847175714/work
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1721585709575/work
pycodestyle==2.12.1
pyflakes==3.2.0
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1714846767233/work
pyparsing @ file:///C:/b/abs_40z8gyj9wi/croot/pyparsing_1731445739241/work
PyQt5==5.15.10
PyQt5-sip @ file:///C:/b/abs_c0pi2mimq3/croot/pyqt-split_1698769125270/work/pyqt_sip
PySocks @ file:///C:/ci_311/pysocks_1676425991111/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1709299778482/work
pytz==2024.2
pywin32==307
PyYAML @ file:///C:/b/abs_14xkfs39bx/croot/pyyaml_1728657968772/work
pyzmq @ file:///D:/bld/pyzmq_1728642364379/work
requests @ file:///C:/b/abs_c3508vg8ez/croot/requests_1731000584867/work
scikit-learn==1.6.0
scipy @ file:///C:/Users/dev-admin/cr_scipy-release/scipy_1731623378890/work/dist/scipy-1.14.1-cp311-cp311-win_amd64.whl#sha256=b77883327feed73ffd182110a17c07ff5dba284c4917ec19034a8d28ff9cbb45
seaborn==0.13.2
sip @ file:///C:/b/abs_edevan3fce/croot/sip_1698675983372/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.1
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
sympy @ file:///C:/b/abs_4e4p71hdj_/croot/sympy_1724938208509/work
tensorflow-probability==0.25.0
threadpoolctl @ file:///C:/b/abs_def0dwqlft/croot/threadpoolctl_1719407816649/work
tokenize_rt==6.1.0
torch==2.1.0+cu121
torchaudio==2.1.0+cu121
torchvision==0.16.0+cu121
tornado @ file:///D:/bld/tornado_1724956185692/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1713535121073/work
typing_extensions @ file:///C:/b/abs_0as9mdbkfl/croot/typing_extensions_1715268906610/work
tzdata==2024.2
unicodedata2 @ file:///C:/b/abs_b6apldlg7y/croot/unicodedata2_1713212998255/work
urllib3 @ file:///C:/b/abs_9a_f8h_bn2/croot/urllib3_1727769836930/work
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
win-inet-pton @ file:///C:/ci_311/win_inet_pton_1676425458225/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1726248574750/work